requirements.txt
fastapi
uvicorn
pydantic
numpy
scipy
pytest
llama-index-llms-cohere
llama-index-embeddings-cohere
psycopg2-binary